+++ tests/__init__.py
@@ -1,1 +1,4 @@
-"""Helix Platform Doctor test suite."""
+# Copyright (c) 2026 Samsarix LLC
+# SPDX-License-Identifier: MPL-2.0
+
+"""Samsarix Platform Doctor test suite."""